requirements.txt
beautifulsoup4==4.6.0
bleach==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.7
django-contrib-comments==1.8.0
filebrowser-safe==0.4.7
future==0.16.0
gitdb2==2.0.2
GitPython==2.1.5
grappelli-safe==0.4.6
html5lib==0.999999999
idna==2.5
Mezzanine==4.2.3
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
requests==2.18.3
requests-oauthlib==0.8.0
rollbar==0.13.12
six==1.10.0
smmap2==2.0.3
tzlocal==1.4
urllib3==1.22
webencodings==0.5.1
